JIPI-8 use Hibernate 4.3 JtaPlatformProvider service to set default JTA platform and eliminate core module
