--- core/pom.xml
@@ -1,100 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<!--
-  ~  JBoss, Home of Professional Open Source
-  ~  Copyright 2013, Red Hat Middleware LLC, and individual contributors
-  ~  by the @authors tag. See the copyright.txt in the distribution for a
-  ~  full listing of individual contributors.
-  ~
-  ~  Licensed under the Apache License, Version 2.0 (the "License");
-  ~  you may not use this file except in compliance with the License.
-  ~  You may obtain a copy of the License at
-  ~  http://www.apache.org/licenses/LICENSE-2.0
-  ~  Unless required by applicable law or agreed to in writing, software
-  ~  distributed under the License is distributed on an "AS IS" BASIS,
-  ~  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~  See the License for the specific language governing permissions and
-  ~  limitations under the License.
-  -->
-
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-
-    <parent>
-        <groupId>org.jipijapa</groupId>
-        <artifactId>jipijapa-parent</artifactId>
-        <version>1.0.0.Alpha2-SNAPSHOT</version>
-    </parent>
-
-    <groupId>org.jipijapa</groupId>
-    <artifactId>core</artifactId>
-    <version>1.0.0.Alpha2-SNAPSHOT</version>
-
-    <name>jipijapa CORE</name>
-
-    <dependencies>
-        <!-- Internal -->
-        <dependency>
-            <groupId>org.jipijapa</groupId>
-            <artifactId>spi</artifactId>
-            <version>${project.version}</version>
-            <scope>provided</scope>
-        </dependency>
-
-        <!-- External -->
-
-        <dependency>
-            <groupId>org.jboss.logging</groupId>
-            <artifactId>jboss-logging</artifactId>
-            <version>${version.org.jboss.logging.jboss-logging}</version>
-            <scope>provided</scope>
-        </dependency>
-
-        <dependency>
-            <groupId>org.jboss.logging</groupId>
-            <artifactId>jboss-logging-processor</artifactId>
-            <version>${version.org.jboss.logging.jboss-logging-processor}</version>
-            <scope>provided</scope>
-        </dependency>
-
-        <dependency>
-            <groupId>org.jboss.logging</groupId>
-            <artifactId>jul-to-slf4j-stub</artifactId>
-            <version>${version.org.jboss.logging.jul-to-slf4j-stub}</version>
-            <scope>provided</scope>
-        </dependency>
-
-        <dependency>
-            <groupId>org.jboss.logmanager</groupId>
-            <artifactId>jboss-logmanager</artifactId>
-            <version>${version.org.jboss.logmanager.jboss-logmanager}</version>
-            <scope>provided</scope>
-        </dependency>
-
-        <dependency>
-            <groupId>org.jboss.logmanager</groupId>
-            <artifactId>log4j-jboss-logmanager</artifactId>
-            <version>${version.org.jboss.logmanager.log4j-jboss-logmanager}</version>
-            <scope>provided</scope>
-        </dependency>
-
-        <dependency>
-            <groupId>org.jboss.slf4j</groupId>
-            <artifactId>slf4j-jboss-logmanager</artifactId>
-            <version>${version.org.jboss.slf4j.slf4j-jboss-logmanager}</version>
-            <scope>provided</scope>
-        </dependency>
-
-        <dependency>
-            <groupId>org.slf4j</groupId>
-            <artifactId>slf4j-api</artifactId>
-            <version>${version.org.slf4j}</version>
-            <scope>provided</scope>
-        </dependency>
-
-
-    </dependencies>
-
-</project>
