+++ eclipselink/pom.xml
@@ -42,12 +42,6 @@
             <scope>provided</scope>
         </dependency>
 
-        <dependency>
-            <groupId>org.jipijapa</groupId>
-            <artifactId>core</artifactId>
-            <scope>provided</scope>
-        </dependency>
-
         <!-- External -->
 
         <dependency>
