+++ hibernate3/src/main/java/org/jboss/as/jpa/hibernate3/HibernatePatternFilter.java
@@ -22,7 +22,7 @@
 
 package org.jboss.as.jpa.hibernate3;
 
-import static org.jipijapa.core.JpaMessages.MESSAGES;
+import static org.jboss.as.jpa.hibernate3.JpaMessages.MESSAGES;
 
 import org.jboss.vfs.VirtualFile;
 import org.jboss.vfs.VirtualFileFilter;
