+++ hibernate4_1/src/main/java/org/jboss/as/jpa/hibernate4/HibernateAnnotationScanner.java
@@ -17,8 +17,8 @@
 
 package org.jboss.as.jpa.hibernate4;
 
-import static org.jipijapa.core.JpaLogger.JPA_LOGGER;
-import static org.jipijapa.core.JpaMessages.MESSAGES;
+import static org.jboss.as.jpa.hibernate4.JpaLogger.JPA_LOGGER;
+import static org.jboss.as.jpa.hibernate4.JpaMessages.MESSAGES;
 
 import java.io.IOException;
 import java.lang.annotation.Annotation;
