+++ hibernate4_1/src/main/java/org/jboss/as/jpa/hibernate4/HibernateLazyNamedInputStream.java
@@ -17,7 +17,7 @@
 
 package org.jboss.as.jpa.hibernate4;
 
-import static org.jipijapa.core.JpaMessages.MESSAGES;
+import static org.jboss.as.jpa.hibernate4.JpaMessages.MESSAGES;
 
 import java.io.IOException;
 import java.io.InputStream;
