+++ hibernate4_3/pom.xml
@@ -65,12 +65,6 @@
             <scope>provided</scope>
         </dependency>
 
-        <dependency>
-            <groupId>org.jipijapa</groupId>
-            <artifactId>core</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-
         <!-- External -->
 
         <dependency>
