+++ openjpa/pom.xml
@@ -43,12 +43,6 @@
             <artifactId>spi</artifactId>
         </dependency>
 
-        <dependency>
-            <groupId>org.jipijapa</groupId>
-            <artifactId>core</artifactId>
-            <scope>provided</scope>
-        </dependency>
-
         <!-- External -->
 
         <dependency>
